@@ -30,8 +30,8 @@ allprojects {
 
     ext {
         minSdk = 17
-        targetSdk = 32
-        compileSdk = 32
+        targetSdk = 33
+        compileSdk = 33
         buildTools = '33.0.0'
 
         isReleaseBuild = {
